+++ product-service/handlers/getProductsList/index.js
@@ -1,0 +1,8 @@
+import { products } from "../../mocks";
+
+export const getProductsList = async () => {
+  return {
+    statusCode: 200,
+    body: JSON.stringify(products),
+  };
+};
